Allegro: Fix coordinate scaling and via drill extraction
Internal coordinates are always stored as mils/divisor regardless of
the board units flag. The previous code used different base scales for
metric (1000) vs imperial (25400), but all test boards use metric and
empirical validation against .alg ASCII exports confirms the correct
formula is always 25400/divisor.

Change m_scale from int to double to avoid truncation with large
divisors (e.g. divisor=10000 gives scale=2.54). Add safeScale() to
clamp double-to-int conversions and prevent KiROUND overflow assertions
on boards with large coordinate values.

Fix via drill extraction to use m_Drill directly instead of a
heuristic that rejected valid values. Fix pad SMD detection to use
scaled drill size instead of comparing raw internal units against
a fixed threshold.

Skip 0-byte .brd files in test discovery to handle broken downloads.

+++ pcbnew/pcb_io/allegro/allegro_builder.h
@@ -144,9 +144,9 @@ private:
 
     std::unique_ptr<LAYER_MAPPER> m_layerMapper;
 
-    // The computed scale factor for the board
-    // (based on units and divisor)
-    int m_scale;
+    // Conversion factor from internal units to nanometers.
+    // Internal coordinates are in mils / divisor, so scale = 25400 / divisor.
+    double m_scale;
 };
 
 } // namespace ALLEGRO
